Change the way we detect a fingerprint scanner
With this commit we're able to see if a fingerprint is enrolled
vs just checking if the hardware supports it. The difference is
that if the user does not have a fingerprint set or 'enrolled' then
they don't see the fingerprint prefs.

There's no reason for the user to use the fingerprint prefs if they
don't have a fingerprint set. It just doesn't make sense.

This mirrors the switch for 'Disable QS in the keyguard' where the
user does not see the switch unless the keyguard is secured.

Change-Id: I58b6a9adab108e93e81ba95e74be5604bbe4005e